Refactored a little
