+++ class_as_dict/__init__.py
